requirements.txt
langchain_openai
feedparser
bs4
langchain
langchain_community
faiss-cpu
cohere
requests
os
python-dotenv
